Move schema-test
